resources/requirements.txt
python-dotenv==1.1.0
streamlit==1.55.0
unstructured==0.17.2
tiktoken==0.9.0
langchain-chroma==1.1.0
langchain-groq==1.1.2
langchain-community==0.3.31
protobuf==5.29.4
langchain-huggingface==1.2.0
